+++ src/components/ModalWrapper/ModalWrapper.js
@@ -103,7 +103,6 @@ export default class ModalWrapper extends React.Component {
     } = this.props;
 
     const props = {
-      ...other,
       open: this.state.isOpen,
       onRequestClose: this.handleClose,
       onRequestSubmit: this.handleOnRequestSubmit,
@@ -113,7 +112,7 @@ export default class ModalWrapper extends React.Component {
       ? React.cloneElement(customButton, {
           disabled: disabled,
           onClick: this.handleOpen,
-          inputref: this.triggerButton,
+          inputRef: this.triggerButton,
         })
       : undefined;
 
@@ -126,8 +125,7 @@ export default class ModalWrapper extends React.Component {
             this.handleClose();
             onKeyDown(evt);
           }
-        }}
-      >
+        }}>
         {customButton ? (
           <React.Fragment>{customButtonEl}</React.Fragment>
         ) : (
@@ -137,12 +135,13 @@ export default class ModalWrapper extends React.Component {
             disabled={disabled}
             kind={triggerButtonKind}
             onClick={this.handleOpen}
-            inputref={this.triggerButton}
-          >
+            inputRef={this.triggerButton}>
             {buttonTriggerText}
           </Button>
         )}
-        <Modal {...props}>{children}</Modal>
+        <Modal {...props} {...other}>
+          {children}
+        </Modal>
       </div>
     );
   }