+++ src/types/Navigation/MainNavigationItem.d.ts
@@ -1,15 +1,16 @@
 import * as React from 'react';
 
-declare namespace MainNavigationItem  {
-  interface MainNavigationItemProps extends React.HTMLProps<MainNavigationItem> {
-    children: React.ReactNode
-    subNavigation?: React.ReactNode
-    activeMenuItem?: number | string
-    menuItem?: number | string
-    subNavWideAsContent?: boolean
-    onChangeSub: () => void
+declare namespace MainNavigationItem {
+  interface MainNavigationItemProps
+    extends React.HTMLProps<MainNavigationItem> {
+    children: React.ReactNode;
+    subNavigation?: React.ReactNode;
+    activeMenuItem?: number | string;
+    menuItem?: number | string;
+    subNavWideAsContent?: boolean;
+    onChangeSub?: () => void;
   }
 }
 
-declare class MainNavigationItem extends React.Component<MainNavigationItem.MainNavigationItemProps>{ }
-export = MainNavigationItem
+declare class MainNavigationItem extends React.Component<MainNavigationItem.MainNavigationItemProps> {}
+export = MainNavigationItem;